requirements.txt
streamlit
tensorflow
numpy
pillow
scikit-learn
nltk
pandas
matplotlib
